+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [3.4.0] 2021-04-20
+
+### Added
+
+- Added script to translate elements that use the attribute `data-translate`.
+- Added chronometer.
+- Highlighter for cells with same value
+- Listener for contextmenu (right click) to toggle the pencil
+- Help Button to set a single value in the board
+
+### Fixed
+
+- Input listener was changed to keydown listener to better process the desired value.
+- Input can receive have values from 1-9
+
 ## [3.3.0] 2021-04-20
 
 ### Added